create component and client side api call
